Added testing mode in libreantdb
If we are testing we are not interested
neither in es replicas nor in es shards.

Thus when testing mode is enabled, libreantdb will create and index with just
one shard and no replicas.

Some simple benchmarks had shown that these changes
speed up the tests execution time by about 18%.